Niraj/pid (#20)
* pseudo code

* progress on PID loop

* rebased to origin/master, cleaned makefile

* testing git, not really commiting anything

* attitude library ready for manual control

* restructured magnometer

* testing imu2euler

* rebased makefile

* rebased main to master

* motor fix

* more pwm debugging

* working motorsgit add .git add .

* putting this on master

+++ src/firmware/drivers/adxl345_accel/adxl345.c
@@ -5,11 +5,11 @@
 #include <util/delay.h>
 #include <stdbool.h>
 
-float _a_x;
-float _a_y;
-float _a_z;
+double _a_x;
+double _a_y;
+double _a_z;
 
-void ADXL345_ReadAccel(float *a_x, float *a_y, float *a_z);
+void ADXL345_ReadAccel(double *a_x, double *a_y, double *a_z);
 void ADXL345_Calibrate(void);
 
 void InitializeADXL345(void)
@@ -33,31 +33,31 @@ void InitializeADXL345(void)
     i2c_write(0x80);    // Enable DATA_READY interrupt
     i2c_stop();
     _delay_ms(2);
-    
+
     // Set offset reg to zero
     i2c_writeReg(ADXL345_ADDR, 0x1E, &zero, 1);
     i2c_writeReg(ADXL345_ADDR, 0x1F, &zero, 1);
     i2c_writeReg(ADXL345_ADDR, 0x20, &zero, 1);
-    
+
     _a_x = 0;
     _a_y = 0;
     _a_z = 0;
 
     _delay_ms(2); // Arbitrary delay amount
     ADXL345_Calibrate();
-
 }
+
 void ADXL345_Calibrate(void)
 {
-    float tmpx = 0;
-    float tmpy = 0;
-    float tmpz = 0;
+    double tmpx = 0;
+    double tmpy = 0;
+    double tmpz = 0;
 
     uint8_t a_offx = 0;
     uint8_t a_offy = 0;
     uint8_t a_offz = 0;
 
-    
+
     // take the mean from 10 gyro probes and divide
     // it from the current probe
     for (uint8_t i = 0; i < 10; i++)
@@ -68,13 +68,13 @@ void ADXL345_Calibrate(void)
         tmpz += _a_z;
         _delay_ms(20); // Arbitrary delay amount
     }
-    
-    // Each LSB of output in full-resolution is one-quarter of an 
+
+    // Each LSB of output in full-resolution is one-quarter of an
     // LSB of the offset register. (Datasheet: Offset Calibration)
     a_offx = -round(tmpx / 10 / 4);
     a_offy = -round(tmpy / 10 / 4) ;
     a_offz = -round(((tmpz / 10) - 256) / 4) ;
-    
+
     // Set calibrated offset values
     i2c_writeReg(ADXL345_ADDR, 0x1E, &a_offx, 1);
     i2c_writeReg(ADXL345_ADDR, 0x1F, &a_offy, 1);
@@ -82,9 +82,7 @@ void ADXL345_Calibrate(void)
 
 }
 
-
-
-void ADXL345_ReadAccel(float *a_x, float *a_y, float *a_z)
+void ADXL345_ReadAccel(double *a_x, double *a_y, double *a_z)
 {
     uint8_t buff[6];
 
@@ -94,7 +92,7 @@ void ADXL345_ReadAccel(float *a_x, float *a_y, float *a_z)
 
 
     i2c_receive(ADXL345_ADDR, buff, 6);
-    
+
     *a_x = ((buff[1] << 8) | buff[0]);
     *a_y = ((buff[3] << 8) | buff[2]);
     *a_z = ((buff[5] << 8) | buff[4]);
@@ -103,7 +101,7 @@ void ADXL345_ReadAccel(float *a_x, float *a_y, float *a_z)
 
 void ADXL345_UpdateData(void)
 {
-    float x, y, z;
+    double x, y, z;
     ADXL345_ReadAccel(&x, &y, &z);
 
     _a_x = x;
@@ -111,20 +109,17 @@ void ADXL345_UpdateData(void)
     _a_z = z;
 }
 
-float ADXL345_GetX(void)
+double ADXL345_GetX(void)
 {
     return _a_x / 256;
 }
 
-float ADXL345_GetY(void)
+double ADXL345_GetY(void)
 {
     return _a_y / 256;
 }
 
-float ADXL345_GetZ(void)
+double ADXL345_GetZ(void)
 {
     return _a_z / 256;
 }
-
-
-
