+++ src/firmware/protocols/gyro/gyro.h
@@ -1,8 +1,8 @@
-/* 
-	PAVx -- Pod-Based Autonomous Vehicles 
+/*
+	PAVx -- Pod-Based Autonomous Vehicles
 	Library Created By: Sargis S Yonan
 	March 2017
-*/ 
+*/
 
 #ifndef _GYRO_H_
 #define _GYRO_H_
@@ -13,10 +13,10 @@ void InitializeGyro(void);
 
 void Gyro_Update(void);
 
-float Gyro_GetX(void);
+double Gyro_GetX(void);
 
-float Gyro_GetY(void);
+double Gyro_GetY(void);
 
-float Gyro_GetZ(void);
+double Gyro_GetZ(void);
 
 #endif
